Introduce pivot feature to control partially the inheritance of X,Y coordinates.
Using prctl to set name of threads to see what thread is rendering what frame + chunk.

Add feature offset_frames, so that resuming rendering frames becomes possible.

Do some javascript object trickery so we can reference the project object "script" before the object has been fully declared yet.

+++ src/generator.h
@@ -58,8 +58,8 @@ private:
   std::unordered_map<int64_t, size_t> next_instance_mapping;
   data::settings settings_;
 
-  size_t min_intermediates = 0.;
-  size_t max_intermediates = 30.;
+  int min_intermediates = 0.;
+  int max_intermediates = 30.;
 
   scale_settings scalesettings;
   scene_settings scenesettings;
